Reworked the directory structure of FGO and cleaned some PipelineManager code.
