+++ runtime/regiontemplates/examples/PipelineRTFS-NS-Diff-FGO/CMakeLists.txt
@@ -20,6 +20,8 @@ find_path(NSCALE_BUILD_DIR REQUIRED)
 set(nscalesrcdir ${NSCALE_SRC_DIR})
 set(nscalebuilddir ${NSCALE_BUILD_DIR})
 
+# dependencies for FGO
+#add_subdirectory("fg_reuse")
 
 #includes
 include_directories(${nscalesrcdir}/segment)
@@ -36,10 +38,17 @@ link_directories(${nscalebuilddir}/lib)
 # Create library with components implementation
 set(libname componentnsdifffgo)
 
+# Json lib
+set(libjson json/jsoncpp.cpp)
+
+# Stages lib - For some reason the subdirectory doesn't work (on PipelineManager.cpp:813 the ComponentFactory appear not to be linked with the stages lib)
+#include_directories(${CMAKE_CURRENT_LIST_DIR}/stages)
+#link_directories(${CMAKE_CURRENT_LIST_DIR}/stages)
+#add_subdirectory("stages")
+set(libstages stages/NormalizationComp.cpp stages/Segmentation.cpp stages/FeatureExtraction.cpp stages/TaskFeatures.cpp stages/DiffMaskComp.cpp)
+
 # Source files used by the Runtime System
-set(libcomponentssrcs NormalizationComp.cpp Segmentation.cpp FeatureExtraction.cpp
-         TaskFeatures.cpp DiffMaskComp.cpp jsoncpp.cpp graph/min_cut.cpp graph/reuse_tree.cpp
-         merging.cpp fgm.cpp cutting_algorithms.cpp)
+set(libcomponentssrcs fg_reuse/graph/min_cut.cpp fg_reuse/graph/reuse_tree.cpp fg_reuse/merging.cpp fg_reuse/fgm.cpp fg_reuse/cutting_algorithms.cpp parsing.cpp parsing_helper.cpp)
 
 set(libcomparativeanalysis ../../comparativeanalysis/TaskDiffMask.cpp
             ../../comparativeanalysis/pixelcompare/PixelCompare.cpp)
@@ -48,11 +57,12 @@ set(libcomparativeanalysis ../../comparativeanalysis/TaskDiffMask.cpp
 add_custom_target(copyCacheConfToPipelineRTFS-NS-Diff-FGO COMMAND ${CMAKE_COMMAND} -E copy ${CMAKE_CURRENT_SOURCE_DIR}/cache-conf/rtconf.xml ${CMAKE_CURRENT_BINARY_DIR})
 
 # Add rule to generate the Runtime System, as a shared library (.so)
-add_library("${libname}" SHARED ${libcomponentssrcs} ${libcomparativeanalysis})
+add_library("${libname}" SHARED ${libcomponentssrcs} ${libcomparativeanalysis} ${libjson} ${libstages})
 
 #set_target_properties("${libname}" PROPERTIES LINK_FLAGS "${MPI_LINK_FLAGS}") 
 target_link_libraries("${libname}" runtimesystem)
 target_link_libraries(${libname} regiontemplates)
+#target_link_libraries(${libname} componentnsdifffgostages)
 
 # nscale libs
 target_link_libraries("${libname}" segment normalization featuresAll ${OpenCV_LIBS})
@@ -64,3 +74,4 @@ add_dependencies(PipelineRTFS-NS-Diff-FGO copyCacheConfToPipelineRTFS-NS-Diff-FG
 # Link the executable to the runtime system library
 target_link_libraries(PipelineRTFS-NS-Diff-FGO runtimesystem "${libname}")
 target_link_libraries(PipelineRTFS-NS-Diff-FGO regiontemplates)
+#target_link_libraries(PipelineRTFS-NS-Diff-FGO componentnsdifffgostages)
