+++ runtime/regiontemplates/examples/PipelineRTFS-NS-Diff-FGO/fg_reuse/fgm.cpp
@@ -1,110 +1,5 @@
 #include "fgm.hpp"
 
-void generate_drs(RegionTemplate* rt, 
-	const std::map<int, ArgumentBase*> &expanded_args) {
-
-	// verify every argument
-	for (pair<int, ArgumentBase*> p : expanded_args) {
-		// if an argument is a region template data region
-		if (p.second->getType() == ArgumentBase::RT) {
-			if (((ArgumentRT*)p.second)->isFileInput) {
-				// create the data region and add it to the input region template
-				DenseDataRegion2D *ddr2d = new DenseDataRegion2D();
-				ddr2d->setName(p.second->getName());
-				std::ostringstream oss;
-				oss << p.first;
-				ddr2d->setId(oss.str());
-				ddr2d->setVersion(p.first);
-				ddr2d->setIsAppInput(true);
-				ddr2d->setInputType(DataSourceType::FILE_SYSTEM);
-				ddr2d->setOutputType(DataSourceType::FILE_SYSTEM);
-				ddr2d->setInputFileName(p.second->toString());
-				rt->insertDataRegion(ddr2d);
-			}
-		}
-	}
-}
-
-void generate_drs(RegionTemplate* rt, PipelineComponentBase* stage,
-	const std::map<int, ArgumentBase*> &expanded_args) {
-
-	// verify every argument
-	for (int inp : stage->getInputs()) {
-		// if an argument is a region template data region
-		if (expanded_args.at(inp)->getType() == ArgumentBase::RT) {
-			// if (((ArgumentRT*)expanded_args.at(inp))->isFileInput) {
-				// create the data region and add it to the input region template
-				DenseDataRegion2D *ddr2d = new DenseDataRegion2D();
-				ddr2d->setName(expanded_args.at(inp)->getName());
-				std::ostringstream oss;
-				oss << inp;
-				ddr2d->setId(oss.str());
-				ddr2d->setVersion(inp);
-				ddr2d->setIsAppInput(true);
-				ddr2d->setInputType(DataSourceType::FILE_SYSTEM);
-				ddr2d->setOutputType(DataSourceType::FILE_SYSTEM);
-				ddr2d->setInputFileName(expanded_args.at(inp)->toString());
-				rt->insertDataRegion(ddr2d);
-			// }
-		}
-	}
-}
-
-void add_arguments_to_stages(std::map<int, PipelineComponentBase*> &merged_stages, 
-	std::map<int, ArgumentBase*> &merged_arguments, string name) {
-
-	int i=0;
-	for (pair<int, PipelineComponentBase*>&& stage : merged_stages) {
-		// create the RT isntance
-		RegionTemplate *rt = new RegionTemplate();
-		rt->setName(name);
-
-		// add input arguments to stage, adding them as RT as needed
-		for (int arg_id : stage.second->getInputs()) {
-			ArgumentBase* new_arg = merged_arguments[arg_id]->clone();
-			new_arg->setParent(merged_arguments[arg_id]->getParent());
-			stage.second->addArgument(new_arg);
-			if (new_arg->getType() == ArgumentBase::RT) {
-				// std::cout << "input RT : " << merged_arguments[arg_id]->getName() << std::endl;
-				// insert the region template on the parent stage if the argument is a DR and if the RT wasn't already added
-				if (((RTPipelineComponentBase*)stage.second)->
-						getRegionTemplateInstance(rt->getName()) == NULL) {
-					((RTPipelineComponentBase*)stage.second)->addRegionTemplateInstance(rt, rt->getName());
-				}
-				((RTPipelineComponentBase*)stage.second)->addInputOutputDataRegion(
-					rt->getName(), new_arg->getName(), RTPipelineComponentBase::INPUT);
-			}
-			if (merged_arguments[arg_id]->getParent() != 0) {
-				// verify if the dependency stage was reused
-				int parent = merged_arguments[arg_id]->getParent();
-				// std::cout << "[before]Dependency: " << stage.second->getId() << ":" << stage.second->getName()
-				// 	<< " ->addDependency( " << parent << " )" << std::endl;
-				if (merged_stages[merged_arguments[arg_id]->getParent()]->reused != NULL)
-					parent = merged_stages[merged_arguments[arg_id]->getParent()]->reused->getId();
-				// std::cout << "Dependency: " << stage.second->getId() << ":" << stage.second->getName()
-				// 	<< " ->addDependency( " << parent << " )" << std::endl;
-				((RTPipelineComponentBase*)stage.second)->addDependency(parent);
-			}
-		}
-
-		// add output arguments to stage, adding them as RT as needed
-		for (int arg_id : stage.second->getOutputs()) {
-			ArgumentBase* new_arg = merged_arguments[arg_id]->clone();
-			new_arg->setParent(merged_arguments[arg_id]->getParent());
-			new_arg->setIo(ArgumentBase::output);
-			stage.second->addArgument(new_arg);
-			if (merged_arguments[arg_id]->getType() == ArgumentBase::RT) {
-				// std::cout << "output RT : " << merged_arguments[arg_id]->getName() << std::endl;
-				// insert the region template on the parent stage if the argument is a DR and if the RT wasn't already added
-				if (((RTPipelineComponentBase*)stage.second)->getRegionTemplateInstance(rt->getName()) == NULL)
-					((RTPipelineComponentBase*)stage.second)->addRegionTemplateInstance(rt, rt->getName());
-				((RTPipelineComponentBase*)stage.second)->addInputOutputDataRegion(rt->getName(), 
-					merged_arguments[arg_id]->getName(), RTPipelineComponentBase::OUTPUT);
-			}
-		}
-	}
-}
-
 void fgm::merge_stages_fine_grain(int algorithm, const std::map<int, PipelineComponentBase*> &all_stages, 
 	const std::map<int, PipelineComponentBase*> &stages_ref, std::map<int, PipelineComponentBase*> &merged_stages, 
 	std::map<int, ArgumentBase*> expanded_args, int max_bucket_size, bool shuffle, string dakota_filename) {
