+++ runtime/regiontemplates/examples/PipelineRTFS-NS-Diff-FGO/json/jsoncpp.cpp
@@ -73,7 +73,7 @@ license you like.
 
 
 
-#include "json/json.h"
+#include "json.h"
 
 #ifndef JSON_IS_AMALGAMATION
 #error "Compile with -I PATH_TO_JSON_DIRECTORY"
