--- runtime/regiontemplates/examples/PipelineRTFS-NS-Diff-FGO/old/NormalizationComp.cpp
@@ -1,64 +1,0 @@
-
-#include "NormalizationComp.h"
-
-
-NormalizationComp::NormalizationComp() {
-	this->setComponentName("NormalizationComp");
-	this->addInputOutputDataRegion("tile", "RAW", RTPipelineComponentBase::INPUT);
-	this->addInputOutputDataRegion("tile", "BGR", RTPipelineComponentBase::OUTPUT);
-}
-
-NormalizationComp::~NormalizationComp() {
-
-}
-
-int NormalizationComp::run()
-{
-
-	// Print name and id of the component instance
-	RegionTemplate * inputRt = this->getRegionTemplateInstance("tile");
-	int parameterId = ((ArgumentInt*)this->getArgument(0))->getArgValue();
-	ArgumentFloatArray* args = ((ArgumentFloatArray*)this->getArgument(1));
-
-
-	std::cout << "Executing component: " << this->getComponentName() << " instance id: " << this->getId() <<std::endl;
-
-	if(inputRt != NULL){
-		DenseDataRegion2D *raw = dynamic_cast<DenseDataRegion2D*>(inputRt->getDataRegion("RAW"));
-		if(raw != NULL){
-			// Create output data region
-			DenseDataRegion2D *bgr = new DenseDataRegion2D();
-			// set outpu data region identifier
-			bgr->setName("BGR");
-			bgr->setId(raw->getId());
-			bgr->setVersion(parameterId);
-
-			// insert data region into region template
-			inputRt->insertDataRegion(bgr);
-
-			float targetMean[3];
-			for(int i = 0; i < 3; i++) targetMean[i] = args->getArgValue(i).getArgValue();
-
-			// Create processing task
-			TaskNormalization * normTask = new TaskNormalization(raw, bgr, targetMean);
-
-			this->executeTask(normTask);
-
-		}else{
-			std::cout << __FILE__ << ":" << __LINE__ <<" Data Region is == NULL" << std::endl;
-		}
-
-	}else{
-		std::cout << __FILE__ << ":" << __LINE__ <<" RT == NULL" << std::endl;
-	}
-
-	return 0;
-}
-
-// Create the component factory
-PipelineComponentBase* componentFactoryNorm() {
-	return new NormalizationComp();
-}
-
-// register factory with the runtime system
-bool registeredN = PipelineComponentBase::ComponentFactory::componentRegister("NormalizationComp", &componentFactoryNorm);
