+++ runtime/regiontemplates/examples/PipelineRTFS-NS-Diff-FGO/parsing.cpp
@@ -1,0 +1,440 @@
+#include "parsing.hpp"
+
+/***************************************************************/
+/***************** Arguments parsing functions *****************/
+/***************************************************************/
+
+int find_arg_pos(string s, int argc, char** argv) {
+	for (int i=1; i<argc; i++)
+		if (string(argv[i]).compare(s)==0)
+			return i;
+	return -1;
+}
+
+/***************************************************************/
+/***************** Workflow parsing functions ******************/
+/***************************************************************/
+
+// returns by reference a map of input arguments to an uid from 'workflow_descriptor' with the
+// list if all possible values each argument can get on a separate map, linked with the same uid
+void get_inputs_from_file(FILE* workflow_descriptor, 
+	map<int, ArgumentBase*> &workflow_inputs, 
+	map<int, list<ArgumentBase*>> &parameters_values) {
+
+	char *line = NULL;
+	size_t len = 0;
+
+	// initial ports section beginning and end
+	string ip("<inputPorts>");
+	string ipe("</inputPorts>");
+	
+	// ports section beginning and end
+	string p("<port>");
+	string pe("</port>");
+
+	// argument name, stored before to be consumed when the ArgumentBase type can be set
+	string name;
+
+	// go to the initial ports beginning
+	while (get_line(&line, workflow_descriptor) != -1 && string(line).find(ip) == string::npos)
+		delete line;
+	// cout << "port init begin: " << line << endl;
+
+	// keep getting ports until it reaches the end of initial ports
+	while (get_line(&line, workflow_descriptor) != -1 && string(line).find(ipe) == string::npos) {
+		// consumes the port beginning
+		while (string(line).find(p) == string::npos && get_line(&line, workflow_descriptor) != -1) {
+			delete line;
+			line = NULL;
+		}
+
+		if (line != NULL)
+			delete line;
+		// cout << "port begin: " << line << endl;
+
+		// finds the name field
+		name = get_workflow_name(workflow_descriptor);
+		// cout << "name: " << name << endl;
+
+		// finds the description field
+		string description = get_workflow_field(workflow_descriptor, "text");
+		// cout << "description: " << description << endl;
+
+		// parse the description to get the input value(s)
+		Json::Reader reader;
+		bool wellFormed;
+		Json::Value data;
+
+		wellFormed = reader.parse(description, data, false);
+		if(!wellFormed) {
+			cout << "Failed to parse JSON: " << description << endl << reader.getFormattedErrorMessages() << endl;
+			exit(-3);
+		}
+
+		// create the propper Argument object for each type case also getting
+		// the parameters' values
+		ArgumentBase* inp_arg;
+		list<ArgumentBase*> inp_values;
+		switch (get_port_type(data["type"].asString())) {
+			case parsing::int_t:
+				// create the argument
+				inp_arg = new ArgumentInt();
+				// cout << "int argument: " << name << ", values: " << endl;		
+				// get all possible values for the argument
+				for (int i=0; i<data["values"].size(); i++) {
+					ArgumentBase* val = new ArgumentInt(data["values"][i].asInt());
+					// cout << ((ArgumentInt*)val)->toString() << endl;
+					inp_values.emplace_back(val);
+				}
+				// cout << endl;
+				break;
+			case parsing::string_t:
+				// create the argument
+				inp_arg = new ArgumentString();
+				// cout << "string argument: " << name << ", values: " << endl;		
+				// get all possible values for the argument
+				for (int i=0; i<data["values"].size(); i++) {
+					ArgumentBase* val = new ArgumentString(data["values"][i].asString());
+					// cout << ((ArgumentString*)val)->toString() << endl;
+					inp_values.emplace_back(val);
+				}
+				// cout << endl;
+				break;
+			case parsing::float_t:
+				// create the argument
+				inp_arg = new ArgumentFloat();
+				// cout << "float argument: " << name << ", values: " << endl;		
+				// get all possible values for the argument
+				for (int i=0; i<data["values"].size(); i++) {
+					ArgumentBase* val = new ArgumentFloat(data["values"][i].asFloat());
+					// cout << ((ArgumentFloat*)val)->toString() << endl;
+					inp_values.emplace_back(val);
+				}
+				// cout << endl;
+				break;
+			case parsing::float_array_t:
+				// create the argument
+				inp_arg = new ArgumentFloatArray();
+				// cout << "floatarray argument: " << name << ", values: " << endl;		
+				// get all possible values for the argument
+				for (int i=0; i<data["values"].size(); i++) {
+					ArgumentBase* val = new ArgumentFloatArray();
+					// cout << "[";
+					for (int j=0; j<data["values"][i].size(); j++) {
+						ArgumentFloat temp(data["values"][i][j].asFloat());
+						((ArgumentFloatArray*)val)->addArgValue(temp);
+						// cout << temp.toString() << endl;
+					}
+					// cout << "]";
+					inp_values.emplace_back(val);
+				}
+				// cout << endl;
+				break;
+			case parsing::rt_t:
+				// create the argument
+				inp_arg = new ArgumentRT();
+				// cout << "string argument: " << name << ", values: " << endl;		
+				// get all possible values for the argument
+				for (int i=0; i<data["values"].size(); i++) {
+					ArgumentBase* val = new ArgumentRT(data["values"][i].asString());
+					((ArgumentRT*)val)->isFileInput = true;
+					// cout << ((ArgumentString*)val)->toString() << endl;
+					inp_values.emplace_back(val);
+				}
+				// cout << endl;
+				break;
+			default:
+				exit(-4);
+		}
+
+		// set inp_arg name, id and input type
+		inp_arg->setName(name);
+		int arg_id = new_uid();
+		inp_arg->setId(arg_id);
+		inp_arg->setIo(ArgumentBase::input);
+
+		// add input argument to map
+		workflow_inputs[arg_id] = inp_arg;
+
+		// add list of argument values to map
+		parameters_values[arg_id] = inp_values;
+
+		// consumes the port ending
+		while (get_line(&line, workflow_descriptor) != -1 && string(line).find(pe) == string::npos)
+			delete line;
+			// cout << "not port end: " << line << endl;
+		// cout << "port end: " << line << endl;
+	}
+}
+
+// returns by reference a map of output arguments, mapped by an uid, 'workflow_outputs'
+void get_outputs_from_file(FILE* workflow_descriptor, map<int, ArgumentBase*> &workflow_outputs) {
+
+	char *line = NULL;
+	size_t len = 0;
+
+	// initial ports section beginning and end
+	string ip("<outputPorts>");
+	string ipe("</outputPorts>");
+	
+	// ports section beginning and end
+	string p("<port>");
+	string pe("</port>");
+
+	// go to the initial ports beginning
+	while (get_line(&line, workflow_descriptor) != -1 && string(line).find(ip) == string::npos)
+		delete line;
+	// cout << "port init begin: " << line << endl;
+
+	// keep getting ports until it reaches the end of initial ports
+	while (get_line(&line, workflow_descriptor) != -1 && string(line).find(ipe) == string::npos) {
+		// consumes the port beginning
+		while (string(line).find(p) == string::npos && get_line(&line, workflow_descriptor) != -1) {
+			delete line;
+			line = NULL;
+		}
+		if (line != NULL)
+			delete line;
+		// cout << "port begin: " << line << endl;
+
+		// finds the name field
+		string name = get_workflow_name(workflow_descriptor);
+		// cout << "name: " << name << endl;
+
+		// finds the description field
+		string description = get_workflow_field(workflow_descriptor, "text");
+		// cout << "description: " << description << endl;
+
+		// parse the description to get the input value(s)
+		Json::Reader reader;
+		bool wellFormed;
+		Json::Value data;
+
+		wellFormed = reader.parse(description, data, false);
+		if(!wellFormed) {
+			cout << "Failed to parse JSON: " << description << endl << reader.getFormattedErrorMessages() << endl;
+			exit(-3);
+		}
+
+		// create the propper Argument object for each type case also getting
+		// the parameters' values
+		ArgumentBase* out_arg;
+		switch (get_port_type(data["type"].asString())) {
+			case parsing::int_t:
+				// create the argument
+				out_arg = new ArgumentInt();
+				// cout << "int output: " << name << endl;		
+				break;
+			case parsing::string_t:
+				// create the argument
+				out_arg = new ArgumentString();
+				// cout << "string output: " << name << endl;		
+				break;
+			case parsing::float_t:
+				// create the argument
+				out_arg = new ArgumentFloat();
+				// cout << "float output: " << name << endl;		
+				break;
+			case parsing::float_array_t:
+				// create the argument
+				out_arg = new ArgumentFloatArray();
+				// cout << "floatarray output: " << name << endl;		
+				break;
+			case parsing::rt_t:
+				// create the argument
+				out_arg = new ArgumentRT();
+				// cout << "string output: " << name << endl;		
+				break;
+			default:
+				exit(-4);
+		}
+		
+		out_arg->setName(name);
+		out_arg->setIo(ArgumentBase::output);
+		int new_id = new_uid();
+		out_arg->setId(new_id);
+		workflow_outputs[new_id] = out_arg;
+
+		// consumes the port ending
+		while (get_line(&line, workflow_descriptor) != -1 && string(line).find(pe) == string::npos)
+			delete line;
+			// get_line << "not port end: " << line << endl;
+		// cout << "port end: " << line << endl;
+	}
+	// cout << "port init end: " << line << endl;
+}
+
+// returns by reference the map of stages on its uid, 'base_stages', and also a map
+// of all output arguments of all stages on its uid, 'interstage_arguments'.
+void get_stages_from_file(FILE* workflow_descriptor, 
+	map<int, PipelineComponentBase*> &base_stages, 
+	map<int, ArgumentBase*> &interstage_arguments) {
+
+	char *line = NULL;
+	size_t len = 0;
+
+	string ps("<processors>");
+	string pse("</processors>");
+
+	string p("<processor>");
+	string pe("</processor>");
+
+	// go to the processors beginning
+	while (get_line(&line, workflow_descriptor) != -1 && string(line).find(ps) == string::npos)
+		delete line;
+		// cout << "not processor init begin: " << line << endl;
+	 // cout << "processor init begin: " << line << endl;
+
+	// keep getting single processors until it reaches the end of all processors
+	while (get_line(&line, workflow_descriptor) != -1 && string(line).find(pse) == string::npos) {
+		// consumes the processor beginning
+		while (string(line).find(p) == string::npos && get_line(&line, workflow_descriptor) != -1) {
+			delete line;
+			line = NULL;
+		}
+
+		if (line != NULL)
+			delete line;
+
+		// cout << "processor begin: " << line << endl;
+
+		// get stage fields
+		string name = get_workflow_name(workflow_descriptor);
+		// cout << "name: " << name << endl;
+
+		// get stage command
+		string command = get_workflow_field(workflow_descriptor, "command");
+		// cout << "command: " << command << endl;
+
+		PipelineComponentBase* stage = PipelineComponentBase::ComponentFactory::getComponentFactory(command)();
+		stage->setName(name);
+
+		// workaround to make sure that the RTs, if any, won't leak on the mearging part of the algorithm
+		stage->setLocation(PipelineComponentBase::WORKER_SIDE);
+
+		// get outputs and add them to the map of arguments
+		// list<string> inputs = get_workflow_ports(workflow_descriptor, "inputPorts");
+		// cout << "outputs:" << endl;
+		list<ArgumentBase*> outputs;
+		get_workflow_arguments(workflow_descriptor, outputs);
+		for(list<ArgumentBase*>::iterator i=outputs.begin(); i!=outputs.end(); i++) {
+			// cout << "\t" << (*i)->getId() << ":" << (*i)->getName() << endl;
+			interstage_arguments[(*i)->getId()] = *i;
+			stage->addOutput((*i)->getId());
+		}
+
+		int stg_id = new_uid();
+		// cout << "uid: " << stg_id << endl;
+		// setting task id
+		stage->setId(stg_id);
+		base_stages[stg_id] = stage;
+
+		// consumes the processor ending
+		while (get_line(&line, workflow_descriptor) != -1 && string(line).find(pe) == string::npos)
+			delete line;
+			// cout << "not processor end: " << line << endl;
+		// cout << "processor end: " << line << endl;
+	}
+}
+
+// returns by reference the map of stages to its ids updated. each stage now has
+// the list of input ids
+void connect_stages_from_file(FILE* workflow_descriptor, 
+	map<int, PipelineComponentBase*> &base_stages, 
+	map<int, ArgumentBase*> &interstage_arguments,
+	map<int, ArgumentBase*> &input_arguments,
+	map<int, list<int>> &deps,
+	map<int, ArgumentBase*> &workflow_outputs) {
+
+	char *line = NULL;
+	size_t len = 0;
+
+	string ds("<datalinks>");
+	string dse("</datalinks>");
+
+	string d("<datalink>");
+	string de("</datalink>");
+
+	string sink("<sink type=");
+	string sinke("</sink>");
+
+	string source("<source type=");
+	string sourcee("</source>");
+
+	// go to the datalinks beginning
+	while (get_line(&line, workflow_descriptor) != -1 && string(line).find(ds) == string::npos)
+		delete line;
+		// cout << "not datalink init begin" << line << endl;
+	// cout << "datalink init begin" << line << endl;
+
+	// keep getting single datalinks until it reaches the end of all datalinks
+	while (get_line(&line, workflow_descriptor) != -1 && string(line).find(dse) == string::npos) {
+		// consumes the datalink beginning
+		while (string(line).find(d) == string::npos && get_line(&line, workflow_descriptor) != -1) {
+			delete line;
+			line = NULL;
+		}
+
+		if (line != NULL)
+			delete line;
+
+		// cout << "datalink begin" << line << endl;
+
+		// get sink and source fields
+		vector<general_field_t> all_sink_fields = get_all_fields(workflow_descriptor, sink, sinke);
+		vector<general_field_t> all_source_fields = get_all_fields(workflow_descriptor, source, sourcee);
+
+		// verify if it's a task sink instead of a workflow sink
+		if (all_sink_fields.size() != 1) {
+			// get the sink stage
+			PipelineComponentBase* sink_stg = find_stage(base_stages, all_sink_fields[0].data);
+			// cout << "stage " << sink_stg->getId() << ":" << sink_stg->getName() << " sink" << endl;
+
+			ArgumentBase* arg;
+			// check whether the source is from the workflow arguments or another stage
+			if (all_source_fields.size() == 1) {
+				// if source is workflow argument:
+				arg = find_argument(input_arguments, all_source_fields[0].data);
+				// cout << "source from workflow is " << arg->getId() << ":" << arg->getName() << endl;
+			} else {
+				// if the source is another stage
+				deps[sink_stg->getId()].emplace_back(find_stage(base_stages, all_source_fields[0].data)->getId());
+				arg = find_argument(interstage_arguments, all_source_fields[1].data);
+				arg->setParent(find_stage(base_stages, all_source_fields[0].data)->getId());
+				// cout << "source from stage " << all_source_fields[0].data << " is " << arg->getId() << ":" 
+				// 	<< arg->getName() << " parent " << arg->getParent() << endl;
+			}
+
+			// add the link to the sink stage
+			sink_stg->addInput(arg->getId());
+		} 
+		else {
+			// cout << "workflow argument sink: " << all_sink_fields[0].data << endl;
+			// update workflow output id in order to access it later to retreive the output
+			
+			ArgumentBase* itstg_argument = find_argument(interstage_arguments, all_source_fields[1].data);
+			// cout << "Output " << all_sink_fields[0].data << " connects to argument " << itstg_argument->getId() << 
+			// 	":" << itstg_argument->getName() << endl;
+			ArgumentBase* output = find_argument(workflow_outputs, all_sink_fields[0].data);
+			// cout << "Output " << all_sink_fields[0].data << " had id " << output->getId() << endl;
+			
+			// remove reference of old id from map
+			workflow_outputs.erase(output->getId());
+
+			// update the output id
+			output->setId(itstg_argument->getId());
+			// cout << "Output " << all_sink_fields[0].data << " now has id " << output->getId() << endl;
+			
+			// re-insert the output with the new id
+			workflow_outputs[output->getId()] = output;
+		}
+
+		// consumes the datalink ending
+		bool cond = string(line).find(de) == string::npos;
+		while (cond && get_line(&line, workflow_descriptor) != -1) {
+			cond = string(line).find(de) == string::npos;
+			delete line;
+		}
+		// cout << "datalink end" << line << endl;
+	}
+}
