+++ runtime/regiontemplates/examples/PipelineRTFS-NS-Diff-FGO/parsing_helper.cpp
@@ -1,0 +1,227 @@
+#include "parsing.hpp"
+
+/***************************************************************/
+/************* Workflow parsing helper functions ***************/
+/***************************************************************/
+
+int get_line(char** line, FILE* f) {
+	char* nline;
+	size_t length=0;
+	if (line_buffer.empty()) {
+		if (getline(&nline, &length, f) == -1)
+			return -1;
+		string sline(nline);
+		size_t pos=string::npos;
+		while ((pos = sline.find("><")) != string::npos) {
+			line_buffer.emplace_back(sline.substr(0,pos+1));
+			sline = sline.substr(pos+1);
+		}
+		line_buffer.emplace_back(sline);
+	}
+
+	char* cline = (char*)malloc((line_buffer.front().length()+1)*sizeof(char*));
+	memcpy(cline, line_buffer.front().c_str(), line_buffer.front().length()+1);
+	*line = cline;
+	line_buffer.pop_front();
+	return strlen(*line);
+}
+
+string get_workflow_name(FILE* workflow) {
+	return get_workflow_field(workflow, "name");
+}
+
+string get_workflow_field(FILE* workflow, string field) {
+	char *line = NULL;
+	size_t len = 0;
+	
+	// create field regex
+	regex r ("<" + field + ">[\"\\:\\w {},.~\\/\\[\\]-]+<\\/" + field + ">");
+
+	// get a new line until name is found
+	while (get_line(&line, workflow) != -1) {
+		smatch match;
+		string s(line);
+		regex_search(s, match, r);
+
+		// cout << "line: " << s << endl;
+
+		// if got a name match
+		if (match.size() == 1) {
+			// cout << "field match: " << line << endl;
+			return s.substr(s.find("<" + field + ">")+field.length()+2, 
+				s.find("</" + field + ">")-s.find("<" + field + ">")-field.length()-2);
+		}
+		delete line;
+	}
+
+	return nullptr;
+}
+
+void get_workflow_arguments(FILE* workflow, 
+	list<ArgumentBase*> &output_arguments) {
+
+	char *line = NULL;
+	size_t len = 0;
+
+	// initial ports section beginning and end
+	string ie("<outputs>");
+	string iee("</outputs>");
+	
+	// ports section beginning and end
+	string e("<entry>");
+	string ee("</entry>");
+
+	// go to the initial entries beginning
+	while (get_line(&line, workflow) != -1 && string(line).find(ie) == string::npos) {
+		delete line;
+		line = NULL;
+	}
+	// cout << "port init begin: " << line << endl;
+
+	// keep getting ports until it reaches the end of initial ports
+	while (get_line(&line, workflow) != -1 && string(line).find(iee) == string::npos) {
+		// consumes the port beginning
+		while (string(line).find(e) == string::npos && get_line(&line, workflow) != -1)
+			delete line;
+		// cout << "port begin: " << line << endl;
+
+		// finds the name and field
+		string name = get_workflow_field(workflow, "string");
+
+		// generate an argument
+		string type = get_workflow_field(workflow, "path");
+		ArgumentBase* arg = new_typed_arg_base(type);
+		arg->setName(name);
+		arg->setId(new_uid());
+		output_arguments.emplace_back(arg);
+
+		// consumes the port ending
+		while (get_line(&line, workflow) != -1 && string(line).find(ee) == string::npos) {
+			delete line;
+			// some c++ implementations may not set line to null after delete
+			line = NULL;
+		}
+		// 	cout << "not port end: " << line << endl;
+		// cout << "port end: " << line << endl;
+
+		if (line != NULL) {
+			delete line;
+			line = NULL;
+		}
+
+	}
+
+	if (line != NULL)
+		delete line;
+	// cout << "port init end: " << line << endl;
+}
+
+vector<general_field_t> get_all_fields(FILE* workflow, string start, string end) {
+	char *line = NULL;
+	size_t len = 0;
+	string type;
+	string field;
+	vector<general_field_t> fields;
+	general_field_t general_field;
+
+	// consumes the beginning
+	while (get_line(&line, workflow) != -1 && string(line).find(start) == string::npos)
+		delete line;
+
+	// create general field regex
+	regex r ("<[\\w]+>[\\w ]+<\\/[\\w]+>");
+	
+	// keep fiding fields until the end
+	while (get_line(&line, workflow) != -1 && string(line).find(end) == string::npos) {
+		smatch match;
+		string s(line);
+		regex_search(s, match, r);
+
+		// if got a general field match
+		if (match.size() == 1) {
+			// cout << "general field match: " << line << endl;
+			type = s.substr(s.find("<")+1, s.find(">")-s.find("<")-1);
+			field = s.substr(s.find("<" + type + ">")+type.length()+2, s.find("</" + type + ">")-s.find("<" + type + ">")-type.length()-2);
+			// cout << "type: " << type << ", field: " << field << endl;
+			general_field.type = type;
+			general_field.data = field;
+			fields.push_back(general_field);
+		}
+		delete line;
+	}
+	return fields;
+}
+
+PipelineComponentBase* find_stage(map<int, PipelineComponentBase*> stages, string name) {
+	for (pair<int, PipelineComponentBase*> p : stages)
+		if (p.second->getName().compare(name) == 0)
+			return p.second;
+	return NULL;
+}
+
+int find_stage_id(map<int, PipelineComponentBase*> stages, string name) {
+	for (pair<int, PipelineComponentBase*> p : stages)
+		if (p.second->getName().compare(name) == 0)
+			return p.first;
+	return -1;
+}
+
+ArgumentBase* find_argument(const map<int, ArgumentBase*>& arguments, string name) {
+	for (pair<int, ArgumentBase*> p : arguments)
+		if (p.second->getName().compare(name) == 0)
+			return p.second;
+	return NULL;
+}
+
+ArgumentBase* find_argument(const list<ArgumentBase*>& arguments, int id) {
+	for (ArgumentBase* a : arguments)
+		if (a->getId() == id)
+			return a;
+	return NULL;
+}
+
+ArgumentBase* find_argument(const list<ArgumentBase*>& arguments, string name) {
+	for (ArgumentBase* a : arguments)
+		if (a->getName().compare(name) == 0)
+			return a;
+	return NULL;
+}
+
+// taken from: http://stackoverflow.com/questions/16388510/evaluate-a-string-with-a-switch-in-c
+constexpr unsigned int str2int(const char* str, int h = 0) {
+	return !str[h] ? 5381 : (str2int(str, h+1) * 33) ^ str[h];
+}
+
+ArgumentBase* new_typed_arg_base(string type) {
+	switch (str2int(type.c_str())) {
+		case str2int("integer"):
+			return new ArgumentInt();
+		case str2int("float"):
+			return new ArgumentFloat();
+		case str2int("string"):
+			return new ArgumentString();
+		case str2int("floatarray"):
+			return new ArgumentFloatArray();
+		case str2int("rt"):
+			return new ArgumentRT();
+		default:
+			return NULL;
+	}
+}
+
+parsing::port_type_t get_port_type(string s) {
+	switch (str2int(s.c_str())) {
+		case str2int("integer"):
+			return parsing::int_t;
+		case str2int("float"):
+			return parsing::float_t;
+		case str2int("string"):
+			return parsing::string_t;
+		case str2int("floatarray"):
+			return parsing::float_array_t;
+		case str2int("rt"):
+			return parsing::rt_t;
+		default:
+			return parsing::error;
+	}
+}
