+++ runtime/regiontemplates/examples/PipelineRTFS-NS-Diff-FGO/stages/DiffMaskComp.cpp
