+++ runtime/regiontemplates/examples/PipelineRTFS-NS-Diff-FGO/stages/FeatureExtraction.cpp
