+++ runtime/regiontemplates/examples/PipelineRTFS-NS-Diff-FGO/stages/NormalizationComp.cpp
