+++ runtime/regiontemplates/examples/PipelineRTFS-NS-Diff-FGO/stages/Segmentation.cpp
