+++ runtime/regiontemplates/examples/PipelineRTFS-NS-Diff-FGO/stages/TaskFeatures.cpp
